Use portable-atomic which supports multicore critical sections
E.g. for the RP2040 you need to, in the project you are using lilos in,
add:

        cargo add rp2040-hal -F critical-section-impl
        cargo add critical-section

If you compare the two binaries, you see that the difference is for the
rp2040/minimal example, where the lilos atomics have some `dmb sy`
instructions.

This is a reasonably good way to compare executables, run it on this
commit and the prior commit, output to file and diff using your
favourite diff program. Yes the regexp is a bit horrible.

        cargo objdump --release -- -d --no-show-raw-insn |                                       \
          sed                                                                                    \
            -e 's/^[0-9a-fA-F]\+: //'                                                            \
            -e 's/\(0x\)\?[0-9a-fA-F]\+ <\(\S\+\)::h[0-9a-fA-F]\{16\}\(+0x[0-9a-fA-F]\+\)>/\2\3/' \
            -e 's/\(0x\)\?[0-9a-fA-F]\+ <\(\S\+\)::h[0-9a-fA-F]\{16\}>/\2/'                       \
            -e 's/\(0x\)\?[0-9a-fA-F]\+ <\(\S\+\)\(+0x[0-9a-fA-F]\+\)>/\2\3/'                     \
            -e 's/\(0x\)\?[0-9a-fA-F]\+ <\(\S\+\)>/\2/'                                           \
            -e 's/@ imm = #-\?0x[0-9a-fA-F]\+//'